+++ jest.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
+  testTimeout: 30000,
   transform: {
     "^.+\\.tsx?$": ["ts-jest", { useESM: true }]
   },